cnn_captcha_pt_br/requirements.txt
absl-py==0.7.1
astor==0.7.1
bleach==1.5.0
captcha==0.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
easydict==1.8
Flask==1.0.2
gast==0.2.2
grpcio==1.19.0
html5lib==0.9999999
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib
numpy==1.16.2
olefile==0.46
Pillow==4.3.0
protobuf==3.6.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.19.1
six==1.12.0
tensorboard==1.7.0
tensorflow==2.0.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.15.2
